+++ src/main/java/com/github/jbb248/test/PichuSprite.java
@@ -33,7 +33,7 @@ public class PichuSprite extends JBurstSprite
     }
 
     @Override
-    public void update(int elapsed)
+    public void update(double elapsed)
     {
         super.update(elapsed);
